add install cilium on workload cluster

+++ scripts/brutal-aws-cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+echo EXIT.
+echo Read the script and uncomment exit statement below.
 exit 1
 
 # Quick delete all resources that make up the clusters provisioned in this account. (did I say it is brutal)
@@ -8,7 +10,9 @@ exit 1
 set -eoux pipefail
 
 running_instances=$(aws ec2 describe-instances --filters "Name=instance-state-name,Values=running" --query "Reservations[].Instances[].InstanceId" --output=text)
-aws ec2 terminate-instances --instance-ids $running_instances
+if [ -n "$running_instances" ]; then
+  aws ec2 terminate-instances --instance-ids $running_instances
+fi
 
 nat_gateways=$(aws ec2 describe-nat-gateways --query "NatGateways[].NatGatewayId" --output=text)
 for n in ${nat_gateways[@]}; do
@@ -22,6 +26,10 @@ eips=$(aws ec2 describe-addresses --query "Addresses[].AllocationId" --output=te
 for i in ${eips[@]}; do
   aws ec2 release-address --allocation-id $i
 done
+
+echo Done.
+
+echo You may need to delete VPC manually from the console
 
 # Some time after NAT and LB are deleted and ENIs are deleted it will be possible to simply delete VPC from the console.
 # The command will still hang though, if attempted to delete from CLI:
